+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v3.3.0 (2026-03-03)
 
 - ✨ Added `hasValue()` global function which checks if variables are defined and have non-nullable values. If variable inside is "" empty string, 0 int, 0.0 float, nil, {} empty object, [] empty array or false, it returns `false`.
+- ✨ Added `format()` built-in function for strings that allows you to embed values into a string. `{{ "He has %s apples".format(apples.len()) }}`.
 
 ## v3.2.5 (2026-02-22)